Mundo1/MissaoCertificacao/requirements.txt
autopep8==1.6.0
Babel==2.10.3
bleach==5.0.1
bokeh==2.4.3
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cycler==0.11.0
decorator==4.4.2
et-xmlfile==1.1.0
Flask==2.1.3
fonttools==4.34.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
lxml==4.9.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
param==1.12.2
phonenumbers==8.12.51
Pillow==9.1.1
plivo==4.24.1
pycodestyle==2.8.0
pyct==0.4.8
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyviz-comms==2.2.0
PyYAML==6.0
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tkcalendar==1.6.1
tkinter-tooltip==2.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
validate-docbr==1.10.0
webencodings==0.5.1
Werkzeug==2.1.2
